backend/requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt>=5.2
django-cors-headers==3.7.0
requests>=2.27.1
python-dotenv==1.0.0
psycopg2-binary>=2.8